Reverted dev changes
The damage of flamethrower was 50 now I reverted it back to 2 as well as a typo in a line for checking weither the w key is pressed for the menu navigation. I reverted the visibility of the collision blocks to false so that is better for the player experience. I added a reset for stamina on game restart as well.