Stamina reasonable and bugfixed
Now the player will lose 10 stamina points for the first attack, and for the combos, he loses 5 for each. Also the regen for stamina is better now. Instead of listening for a keypress only it now also checks wether the stamina is over 10 to regen, and when it does regen, it regens faster.